treewide: plug memleak: case 2.2.5: of_for_each_phandle() (v6.12 prep)
Case 2.2.5 from below bug:

    If returning or breaking early from within of_for_each_phandle(), a
    reference to current iterator node remains and needs to be dropped,
    either right away, or later if used past that loop to avoid a
    memleak.

In these cases here, the np is not needed past the loop, and therefore
the extra reference bumps need to be removed (of_for_each_phandle()
holds a reference to the iterator node already). Care must also be
taken to drop the reference that otherwise remains when exiting the
loop early (via break or goto).

Do so.

This commit plugs memleaks in both normal and error code paths.

Test: TH
Bug: 370679843
Change-Id: I38d29d34c21269d64b25244c34b79e01ed4ba517
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>